Refactored: To Retrieve Employee Who Joined In A Particular Data Range

+++ src/com/bridgelab/EmployeePayrollData.java
@@ -4,23 +4,28 @@ import java.sql.Date;
 
 public class EmployeePayrollData {
 	private int id;
-	private String name;
-	private String gender;
-	private int salary;
-	private Date start_date;
+    private String name;
+    private String gender;
+    private double salary;
+    private Date start_date;
 
-	public EmployeePayrollData(int id, String name, String gender, int salary, Date start_date) {
-		this.id = id;
-		this.name = name;
-		this.gender = gender;
-		this.salary = salary;
-		this.start_date = start_date;
-	}
+    public EmployeePayrollData(int id, String name, String gender, double salary, Date start_date) {
+        this.id = id;
+        this.name = name;
+        this.gender = gender;
+        this.salary = salary;
+        this.start_date = start_date;
+    }
 
-	@Override
-	public String toString() {
-		return "EmployeePayrollData{" + "id=" + id + ", name='" + name + '\'' + ", gender='" + gender + '\''
-				+ ", salary=" + salary + ", start_date=" + start_date + '}';
-	}
+    @Override
+    public String toString() {
+        return "EmployeePayrollData{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", gender='" + gender + '\'' +
+                ", salary=" + salary +
+                ", start_date=" + start_date +
+                '}';
+    }
 
 }
